+++ packages/medusa-fulfillment-webshipper/src/services/webshipper-fulfillment.js
@@ -9,6 +9,16 @@ class WebshipperFulfillmentService extends FulfillmentService {
 
     this.options_ = options
 
+    if (!options.coo_countries) {
+      this.options_.coo_countries = ["all"]
+    } else if (Array.isArray(options.coo_countries)) {
+      this.options_.coo_countries = options.coo_countries.map((c) =>
+        c.toLowerCase()
+      )
+    } else if (typeof options.coo_countries === "string") {
+      this.options_.coo_countries = [options.coo_countries]
+    }
+
     /** @private @const {logger} */
     this.logger_ = logger
 
@@ -242,8 +252,12 @@ class WebshipperFulfillmentService extends FulfillmentService {
       webshipperOrder = await this.client_.orders.retrieve(existing)
     }
 
+    const { shipping_address } = fromOrder
+
     if (!webshipperOrder) {
       let invoice
+      let certificateOfOrigin
+
       if (this.invoiceGenerator_) {
         const base64Invoice = await this.invoiceGenerator_.createInvoice(
           fromOrder,
@@ -263,6 +277,34 @@ class WebshipperFulfillmentService extends FulfillmentService {
           .catch((err) => {
             throw err
           })
+
+        const cooCountries = this.options_.coo_countries
+        if (
+          (cooCountries.includes("all") ||
+            cooCountries.includes(
+              shipping_address.country_code.toLowerCase()
+            )) &&
+          this.invoiceGenerator_.createCertificateOfOrigin
+        ) {
+          const base64Coo = await this.invoiceGenerator_.createCertificateOfOrigin(
+            fromOrder,
+            fulfillmentItems
+          )
+
+          certificateOfOrigin = await this.client_.documents
+            .create({
+              type: "documents",
+              attributes: {
+                document_size: this.options_.document_size || "A4",
+                document_format: "PDF",
+                base64: base64Coo,
+                document_type: "certificate",
+              },
+            })
+            .catch((err) => {
+              throw err
+            })
+        }
       }
 
       let id = fulfillment.id
@@ -274,7 +316,6 @@ class WebshipperFulfillmentService extends FulfillmentService {
         visible_ref = `S-${fromOrder.display_id}`
       }
 
-      const { shipping_address } = fromOrder
       const newOrder = {
         type: "orders",
         attributes: {
@@ -335,14 +376,25 @@ class WebshipperFulfillmentService extends FulfillmentService {
           country_code: methodData.drop_point_country_code.toUpperCase(),
         }
       }
-      if (invoice) {
+
+      if (invoice || certificateOfOrigin) {
+        const docData = []
+        if (invoice) {
+          docData.push({
+            id: invoice.data.id,
+            type: invoice.data.type,
+          })
+        }
+
+        if (certificateOfOrigin) {
+          docData.push({
+            id: certificateOfOrigin.data.id,
+            type: certificateOfOrigin.data.type,
+          })
+        }
+
         newOrder.relationships.documents = {
-          data: [
-            {
-              id: invoice.data.id,
-              type: invoice.data.type,
-            },
-          ],
+          data: docData,
         }
       }