+++ packages/medusa/src/api/routes/admin/orders/list-orders.js
@@ -6,7 +6,7 @@ import { MedusaError, Validator } from "medusa-core-utils"
  * @oas [get] /orders
  * operationId: "GetOrders"
  * summary: "List Orders"
- * description: "Retrieves an list of Orders"
+ * description: "Retrieves a list of Orders"
  * tags:
  *   - Order
  * responses:
@@ -16,8 +16,10 @@ import { MedusaError, Validator } from "medusa-core-utils"
  *       application/json:
  *         schema:
  *           properties:
- *             order:
- *               $ref: "#/components/schemas/order"
+ *             orders:
+ *               type: array
+ *               items:
+ *                 $ref: "#/components/schemas/order"
  */
 export default async (req, res) => {
   const schema = Validator.orderFilter()